“To predict the possibility if the client will deposit or not” (Data mining project using python)
The objective is to examine the performance of MLPNN, TAN, LR etc. techniques on a real-world data of bank deposit subscription. The purpose is increasing the campaign effectiveness by identifying the main characteristics that affect a success (the deposit subscribed by the client) based on MLPNN, TAN, LR. The experimental results demonstrate, with higher accuracies, the success of these models in predicting the best campaign contact with the clients for subscribing deposit.